Fix error getters
