+++ canvas_modules/common-canvas/src/common-canvas/svg-canvas-utils-nodes.js
@@ -15,7 +15,8 @@
  */
 
 import CanvasUtils from "./common-canvas-utils.js";
-import { SUPER_NODE, TEXT_AREA_BORDER_ADJUSTMENT } from "./constants/canvas-constants";
+import { ERROR, WARNING, NODE_ERROR_ICON, NODE_WARNING_ICON,
+	SUPER_NODE, TEXT_AREA_BORDER_ADJUSTMENT } from "./constants/canvas-constants";
 
 export default class SvgCanvasNodes {
 	constructor(config, canvasLayout) {
@@ -48,6 +49,68 @@ export default class SvgCanvasNodes {
 		const lineTypeClass = node.layout.labelSingleLine ? " d3-label-single-line" : " d3-label-multi-line";
 		const justificationClass = node.layout.labelAlign === "center" ? " d3-label-center" : "";
 		return "d3-node-label-entry" + lineTypeClass + justificationClass;
+	}
+
+	getErrorMarkerIcon(data) {
+		const messageLevel = this.getMessageLevel(data.messages);
+		let iconPath = "";
+		switch (messageLevel) {
+		case ERROR:
+			iconPath = NODE_ERROR_ICON;
+			break;
+		case WARNING:
+			iconPath = NODE_WARNING_ICON;
+			break;
+		default:
+			break;
+		}
+		return iconPath;
+	}
+
+	getMessageLevel(messages) {
+		let messageLevel = "";
+		if (messages && messages.length > 0) {
+			for (const message of messages) {
+				if (message.type === ERROR) {
+					return message.type;
+				} else if (message.type === WARNING) {
+					messageLevel = message.type;
+				}
+			}
+		}
+		return messageLevel;
+	}
+
+	getMessageLabelClass(messages) {
+		const messageLevel = this.getMessageLevel(messages);
+		let labelClass = "";
+		switch (messageLevel) {
+		case ERROR:
+			labelClass = "d3-node-error-label";
+			break;
+		case WARNING:
+			labelClass = "d3-node-warning-label";
+			break;
+		default:
+			break;
+		}
+		return labelClass;
+	}
+
+	getErrorMarkerClass(messages) {
+		const messageLevel = this.getMessageLevel(messages);
+		let labelClass = "d3-error-circle-off";
+		switch (messageLevel) {
+		case ERROR:
+			labelClass = "d3-error-circle";
+			break;
+		case WARNING:
+			labelClass = "d3-warning-circle";
+			break;
+		default:
+			break;
+		}
+		return labelClass;
 	}
 
 	getNodeImagePosX(node) {
